+++ src/rules/validation/abstract/base/RuleBlacklistBase.sol
@@ -5,6 +5,7 @@ import {RuleAddressSet} from "../RuleAddressSet/RuleAddressSet.sol";
 import {RuleNFTAdapter} from "../core/RuleNFTAdapter.sol";
 import {RuleTransferValidation} from "../core/RuleTransferValidation.sol";
 import {RuleBlacklistInvariantStorage} from "../RuleAddressSet/invariantStorage/RuleBlacklistInvariantStorage.sol";
+import {AddressListInterfaceId} from "../../../interfaces/library/AddressListInterfaceId.sol";
 import {IERC1404, IERC1404Extend} from "CMTAT/interfaces/tokenization/draft-IERC1404.sol";
 import {IERC3643IComplianceContract} from "CMTAT/interfaces/tokenization/IERC3643Partial.sol";
 import {IRuleEngine} from "CMTAT/interfaces/engine/IRuleEngine.sol";
@@ -94,7 +95,10 @@ abstract contract RuleBlacklistBase is RuleAddressSet, RuleNFTAdapter, RuleBlack
      * @inheritdoc RuleTransferValidation
      */
     function supportsInterface(bytes4 interfaceId) public view virtual override(RuleTransferValidation) returns (bool) {
-        return RuleTransferValidation.supportsInterface(interfaceId);
+        // Advertise IAddressList: this rule manages an address set and is callable through
+        // the IAddressList interface.
+        return interfaceId == AddressListInterfaceId.IADDRESS_LIST_INTERFACE_ID
+            || RuleTransferValidation.supportsInterface(interfaceId);
     }
 
     /*//////////////////////////////////////////////////////////////